Don't actually need to have separate trait and actor for ClearingMechanismActor.
