+++ src/main/scala-2.11/markets/clearing/ClearingMechanismActor.scala
@@ -17,19 +17,44 @@ package markets.clearing
 
 import akka.actor.{Props, ActorRef, Actor}
 import markets.clearing.engines.MatchingEngineLike
+import markets.orders.OrderLike
+
+import scala.util.{Failure, Success}
 
 
-/** Base implementation of a `ClearingMechanismActor`.
+/** Actor for modeling market clearing mechanisms.
   *
-  * @note Users should always be able to create a `ClearingMechanismLike` actor directly using this class. For
-  *       convenience, many standard clearing mechanisms should inherit from this class.
+  * A `ClearingMechanismActor` actor should receive orders and fill them using its matching engine. Filled
+  * orders are then sent to a `SettlementMechanismLike` actor for further processing.
+  * @param matchingEngine A `ClearingMechanismActor` has a matching engine for forming prices and quantities
+  * @param settlementMechanism A `ClearingMechanismActor` has access to some settlement mechanism that it uses to
+  *                            process filled orders into successful transactions
   */
 class ClearingMechanismActor(val matchingEngine: MatchingEngineLike,
-                             val settlementMechanism: ActorRef) extends Actor with ClearingMechanismLike
+                             val settlementMechanism: ActorRef) extends Actor {
+
+  def receive: Receive = {
+    case order: OrderLike =>
+      val result = matchingEngine.fillIncomingOrder(order)
+      result match {
+        case Success(filledOrders) =>
+          filledOrders.foreach(filledOrder => settlementMechanism ! filledOrder)
+        case Failure(ex) =>
+          order.issuer ! ex
+      }
+    case _ => ???
+
+  }
+
+}
 
 
 /** Companion object for `ClearingMechanismActor`. */
 object ClearingMechanismActor {
+
+  def apply(matchingEngine: MatchingEngineLike, settlementMechanism: ActorRef): ClearingMechanismActor = {
+    new ClearingMechanismActor(matchingEngine, settlementMechanism)
+  }
 
   def props(matchingEngine: MatchingEngineLike, settlementMechanism: ActorRef): Props = {
     Props(new ClearingMechanismActor(matchingEngine, settlementMechanism))
